// src/api/generated/schemas/orgUserUpdateRequestRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NervesHub API
 * OpenAPI spec version: 2.0.0
 */

export type OrgUserUpdateRequestRole = typeof OrgUserUpdateRequestRole[keyof typeof OrgUserUpdateRequestRole];


export const OrgUserUpdateRequestRole = {
  admin: 'admin',
  manage: 'manage',
  view: 'view',
} as const;